backend/requirements.txt
# PyRobot Simplified Backend Requirements
# Minimal, focused dependencies for the simplified architecture

# Core FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database access
pyodbc==4.0.39

# Data validation
pydantic==2.5.0

# WebSocket support for real-time updates
websockets==12.0

# Camera and video processing
opencv-python==4.8.1.78

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Environment configuration
python-dotenv==1.0.0

# Windows-specific dependencies
pywin32>=306; sys_platform == "win32"

# System tray icon support (optional)
pystray==0.19.5
Pillow==10.1.0